Created class CommandListener which now houses the onCommand method
